Corrigir registo, envio de email e WhatsApp

+++ backend-sn/components/deleteAdministrador.php
@@ -1,9 +1,15 @@
 <?php
+
+require_once __DIR__ . '/../../vendor/autoload.php';
 require_once '../connect/server.php';
 require_once '../connect/cors.php';
 include './updateAdministradores.php';
 
-$id = $_GET['id_admin'];
-$current_user_id = $_GET['current_user_id']; // Obtenha o ID do usuário atual
-deleteAdmin($id, $current_user_id); // Passe ambos os IDs para a função
+$id = isset($_GET['id_admin']) ? intval($_GET['id_admin']) : 0;
+$current_user_id = isset($_GET['current_user_id']) ? intval($_GET['current_user_id']) : 0;
+if ($id > 0 && $current_user_id > 0) {
+	deleteAdmin($id, $current_user_id);
+} else {
+	echo json_encode(["status" => "error", "message" => "IDs inválidos"]);
+}
 ?>
